[UPDATE] supabase 통신 로직 API로 이동 후 관리
